WSTEAM1-623: Tidy merge
